backend/requirements.txt
fastapi
uvicorn[standard]
asyncpg
sqlalchemy[asyncio]
alembic
python-jose[cryptography]
bcrypt
python-multipart
pydantic-settings
sse-starlette
openai
pgvector
# RAG - extraction de texte
pdfplumber
python-docx
openpyxl
Pillow
pytesseract
# RAG - embeddings
httpx
